Uploaded the data set, the papers and also a small program for reading the files.
Committed:
1) Papers,
2) Data set, and
3) readTweetsFromFile.py, which reads a textfile from the data set and saves the tweets in a dictionary. In the main-function, one can specify which textfile one wants to read. Set the subtask variable to 1 for subtask A, 2 for subtask B and so on.

Each key in the output dictionary is the number of the tweet which is stored in the value of that key. Each value is a list where index 0 is the actual tweet. If subtask A is chosen, index 1 is the sentiment ranking of that tweet. If subtask B or C is chosen, index 1 is the topic and index 2 is the sentiment ranking of that tweet.